move generic data structures to openfpgautil library
