+++ libopenfpga/libopenfpgautil/src/openfpga_port_parser.cpp
@@ -1,50 +1,17 @@
-/**********************************************************
- * MIT License
- *
- * Copyright (c) 2018 LNIS - The University of Utah
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- ***********************************************************************/
-
-/************************************************************************
- * Filename:    port_parser.cpp
- * Created by:   Xifan Tang
- * Change history:
- * +-------------------------------------+
- * |  Date       |    Author   | Notes
- * +-------------------------------------+
- * | 2019/08/09  |  Xifan Tang | Created 
- * +-------------------------------------+
- ***********************************************************************/
-
 /************************************************************************
  * Member functions for Port parsers
  ***********************************************************************/
-
-#include "string.h"
+#include <cstring>
 
 #include "vtr_assert.h"
 #include "vtr_geometry.h"
 
-#include "string_token.h"
+#include "openfpga_tokenizer.h"
 
-#include "port_parser.h"
+#include "openfpga_port_parser.h"
+
+/* namespace openfpga begins */
+namespace openfpga {
 
 /************************************************************************
  * Member functions for PortParser class 
@@ -323,9 +290,4 @@ void PortDelayParser::clear() {
   return;
 }
 
-
-
-/************************************************************************
- * End of file : port_parser.cpp
- ***********************************************************************/
-
+} /* namespace openfpga ends */
